requirements.txt
plyer==2.1.0
pyinstaller==6.11.1
requests==2.32.3
PyYAML==6.0.2
